+++ backend/.serverless/serverless-state.json
@@ -197,7 +197,7 @@
                 "S3Bucket": {
                   "Ref": "ServerlessDeploymentBucket"
                 },
-                "S3Key": "serverless/optimizer/dev/1650490161323-2022-04-20T21:29:21.323Z/optimizer.zip"
+                "S3Key": "serverless/optimizer/dev/1650614958445-2022-04-22T08:09:18.445Z/optimizer.zip"
               },
               "Handler": "api/index.router",
               "Runtime": "nodejs14.x",
@@ -229,7 +229,7 @@
                 "S3Bucket": {
                   "Ref": "ServerlessDeploymentBucket"
                 },
-                "S3Key": "serverless/optimizer/dev/1650490161323-2022-04-20T21:29:21.323Z/optimizer.zip"
+                "S3Key": "serverless/optimizer/dev/1650614958445-2022-04-22T08:09:18.445Z/optimizer.zip"
               },
               "Handler": "processor/index.optimize",
               "Runtime": "nodejs14.x",
@@ -254,24 +254,24 @@
               "OptimizeDashprocessorLogGroup"
             ]
           },
-          "MainLambdaVersion6JoCnKK6jlqNEYsuaLLV6vPZWwc7fIQiom42rDbBoY": {
+          "MainLambdaVersiongpXKw9q1d32YGZOMb832t4jtoePnA8SnXQZUjien2B0": {
             "Type": "AWS::Lambda::Version",
             "DeletionPolicy": "Retain",
             "Properties": {
               "FunctionName": {
                 "Ref": "MainLambdaFunction"
               },
-              "CodeSha256": "qFaD283QhfWxQG80MH9CDsKk96w6aX0DbLPwkO5Qv/Y="
+              "CodeSha256": "i9x9l3leyCS049jBP+LEBURGV8Aem5o62IavacERJTI="
             }
           },
-          "OptimizeDashprocessorLambdaVersionbmTxttOVA3zx0wdoUzTc6jlo4JICiThUzvUcvXks5Y": {
+          "OptimizeDashprocessorLambdaVersionSowLQWu0EOlyE9rbJhFRjO2RXu4s6GNuR0tkudjdowc": {
             "Type": "AWS::Lambda::Version",
             "DeletionPolicy": "Retain",
             "Properties": {
               "FunctionName": {
                 "Ref": "OptimizeDashprocessorLambdaFunction"
               },
-              "CodeSha256": "qFaD283QhfWxQG80MH9CDsKk96w6aX0DbLPwkO5Qv/Y="
+              "CodeSha256": "i9x9l3leyCS049jBP+LEBURGV8Aem5o62IavacERJTI="
             }
           },
           "ApiGatewayRestApi": {
@@ -493,7 +493,7 @@
               "MainLambdaPermissionApiGateway"
             ]
           },
-          "ApiGatewayDeployment1650490159487": {
+          "ApiGatewayDeployment1650614956755": {
             "Type": "AWS::ApiGateway::Deployment",
             "Properties": {
               "RestApiId": {
@@ -558,7 +558,7 @@
           "MainLambdaFunctionQualifiedArn": {
             "Description": "Current Lambda function version",
             "Value": {
-              "Ref": "MainLambdaVersion6JoCnKK6jlqNEYsuaLLV6vPZWwc7fIQiom42rDbBoY"
+              "Ref": "MainLambdaVersiongpXKw9q1d32YGZOMb832t4jtoePnA8SnXQZUjien2B0"
             },
             "Export": {
               "Name": "sls-optimizer-dev-MainLambdaFunctionQualifiedArn"
@@ -567,7 +567,7 @@
           "OptimizeDashprocessorLambdaFunctionQualifiedArn": {
             "Description": "Current Lambda function version",
             "Value": {
-              "Ref": "OptimizeDashprocessorLambdaVersionbmTxttOVA3zx0wdoUzTc6jlo4JICiThUzvUcvXks5Y"
+              "Ref": "OptimizeDashprocessorLambdaVersionSowLQWu0EOlyE9rbJhFRjO2RXu4s6GNuR0tkudjdowc"
             },
             "Export": {
               "Name": "sls-optimizer-dev-OptimizeDashprocessorLambdaFunctionQualifiedArn"
@@ -733,7 +733,7 @@
         "memory": 256,
         "runtime": "nodejs14.x",
         "vpc": {},
-        "versionLogicalId": "MainLambdaVersion6JoCnKK6jlqNEYsuaLLV6vPZWwc7fIQiom42rDbBoY"
+        "versionLogicalId": "MainLambdaVersiongpXKw9q1d32YGZOMb832t4jtoePnA8SnXQZUjien2B0"
       },
       "optimize-processor": {
         "timeout": 900,
@@ -744,7 +744,7 @@
         "memory": 256,
         "runtime": "nodejs14.x",
         "vpc": {},
-        "versionLogicalId": "OptimizeDashprocessorLambdaVersionbmTxttOVA3zx0wdoUzTc6jlo4JICiThUzvUcvXks5Y"
+        "versionLogicalId": "OptimizeDashprocessorLambdaVersionSowLQWu0EOlyE9rbJhFRjO2RXu4s6GNuR0tkudjdowc"
       }
     },
     "configValidationMode": "warn",
@@ -765,7 +765,7 @@
     "artifact": "/home/kav/newrelic/nr1-apps/nr1-cloud-optimize/backend/.serverless/optimizer.zip"
   },
   "package": {
-    "artifactDirectoryName": "serverless/optimizer/dev/1650490161323-2022-04-20T21:29:21.323Z",
+    "artifactDirectoryName": "serverless/optimizer/dev/1650614958445-2022-04-22T08:09:18.445Z",
     "artifact": "optimizer.zip"
   }
 }
