+++ catalog/config.json
@@ -7,7 +7,8 @@
     "gcp",
     "aws",
     "amazon",
-    "alicloud"
+    "alicloud",
+    "optimize"
   ],
   "keywords": [
     "infrastructure"
